Actually use ShapeDragModel
