+++ core/ide/src/main/java/org/alice/stageide/gallerybrowser/shapes/ConeDragModel.java
@@ -42,14 +42,10 @@
  *******************************************************************************/
 package org.alice.stageide.gallerybrowser.shapes;
 
-import edu.cmu.cs.dennisc.math.AxisAlignedBox;
 import org.alice.stageide.ast.declaration.AddConeManagedFieldComposite;
-import org.alice.stageide.icons.ConeIconFactory;
 import org.alice.stageide.modelresource.ResourceNode;
 import org.lgna.croquet.SingleSelectTreeState;
 import org.lgna.croquet.Triggerable;
-import org.lgna.croquet.icon.IconFactory;
-import org.lgna.story.EmployeesOnly;
 import org.lgna.story.SCone;
 
 import java.util.UUID;
@@ -58,8 +54,6 @@ import java.util.UUID;
  * @author Dennis Cosgrove
  */
 public class ConeDragModel extends ShapeDragModel {
-  private static final SCone sModel = new SCone();
-
   private static class SingletonHolder {
     private static ConeDragModel instance = new ConeDragModel();
   }
@@ -69,17 +63,7 @@ public class ConeDragModel extends ShapeDragModel {
   }
 
   private ConeDragModel() {
-    super(UUID.fromString("a3745200-6936-448c-b3c1-fa40c844482a"));
-  }
-
-  @Override
-  public AxisAlignedBox getBoundingBox() {
-    return EmployeesOnly.getImplementation(sModel).getAxisAlignedMinimumBoundingBox();
-  }
-
-  @Override
-  public boolean placeOnGround() {
-    return true;
+    super(UUID.fromString("a3745200-6936-448c-b3c1-fa40c844482a"), new SCone());
   }
 
   @Override
@@ -87,8 +71,5 @@ public class ConeDragModel extends ShapeDragModel {
     return AddConeManagedFieldComposite.getInstance().getLaunchOperation();
   }
 
-  @Override
-  public IconFactory getIconFactory() {
-    return ConeIconFactory.getInstance();
-  }
+
 }
