+++ core/ide/src/main/java/org/alice/stageide/gallerybrowser/shapes/GroundDragModel.java
@@ -42,14 +42,10 @@
  */
 package org.alice.stageide.gallerybrowser.shapes;
 
-import edu.cmu.cs.dennisc.math.AxisAlignedBox;
 import org.alice.stageide.ast.declaration.AddGroundManagedFieldComposite;
-import org.alice.stageide.icons.GroundIconFactory;
 import org.alice.stageide.modelresource.ResourceNode;
 import org.lgna.croquet.SingleSelectTreeState;
 import org.lgna.croquet.Triggerable;
-import org.lgna.croquet.icon.IconFactory;
-import org.lgna.story.EmployeesOnly;
 import org.lgna.story.SGround;
 
 import java.util.UUID;
@@ -58,9 +54,6 @@ import java.util.UUID;
  * @author dculyba
  */
 public class GroundDragModel extends ShapeDragModel {
-
-  private static final SGround sModel = new SGround();
-
   private static class SingletonHolder {
     private static GroundDragModel instance = new GroundDragModel();
   }
@@ -70,27 +63,13 @@ public class GroundDragModel extends ShapeDragModel {
   }
 
   private GroundDragModel() {
-    super(UUID.fromString("3aa18490-f758-46e6-84eb-c9eff76c0185"));
+    super(UUID.fromString("3aa18490-f758-46e6-84eb-c9eff76c0185"), new SGround());
   }
 
-  @Override
-  public AxisAlignedBox getBoundingBox() {
-    return EmployeesOnly.getImplementation(sModel).getAxisAlignedMinimumBoundingBox();
-  }
-
-  @Override
-  public boolean placeOnGround() {
-    return true;
-  }
 
   @Override
   public Triggerable getLeftButtonClickOperation(SingleSelectTreeState<ResourceNode> controller) {
     return AddGroundManagedFieldComposite.getInstance().getLaunchOperation();
-  }
-
-  @Override
-  public IconFactory getIconFactory() {
-    return GroundIconFactory.getInstance();
   }
 
 }
